Cleanup some private utilities in app_commands.commands
