+++ discord/app_commands/transformers.py
@@ -870,8 +870,8 @@ def annotation_to_parameter(annotation: Any, parameter: inspect.Parameter) -> Co
 
     # Check if the method is overridden
     if inner.autocomplete.__func__ is not Transformer.autocomplete:
-        from .commands import _validate_auto_complete_callback
+        from .commands import validate_auto_complete_callback
 
-        result.autocomplete = _validate_auto_complete_callback(inner.autocomplete)
+        result.autocomplete = validate_auto_complete_callback(inner.autocomplete)
 
     return result
